split up source code
