+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "harlowe-macro-api",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "Unofficial Macro API for Twine2/Harlowe.",
   "main": "build.js",
   "dependencies": {},
